+++ readScores.py
