+++ readTimes.py
